Discard Docker build trigger for non-source files
For a successful building of the JAR file, `src` directory and
`pom.xml` are required. However, the `jar` stage tries to copy all
files. This means that if al least file, even non-project, is changed,
the whole build process will be triggered. For preventing of spare
builds, the `jar` stage copies only required source files now.