Wait for Selenium daemon is ready
Sometimes happens that a Selenium daemon (means here webdriver binary
or Selenium Grid) is not ready to work. In this case, the tests should
wait a bit before actually start testing. That's why the waiting logic
is added to verify if the daemon is ready or not for 30 seconds. If
it is not ready, the tests won't start and the execution will fail with
exit code 0.